# swagger.yaml
...
openapi: 3.0.0
info:
  title: Gamery js
  version: "1.0"
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://web-gamery.herokuapp.com
    description: Production
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    body_temporary:
      type: object
      properties:
        uuid:
          type: string
    body_credentials:
      type: object
      properties:
        usernameEmail:
          type: string
        password:
          type: string
    body_registration:
      type: object
      required: true
      properties:
        uuid:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
...
    response_login:
      type: object
      properties:
        data:
          type: object
          properties:
            access:
              type: string
            uuid:
              type: string
    response_full_user:
      type: object
      properties:
        data:
          type: object
          properties:
            _id:
              type: string
            username:
              type: string
            uuid:
              type: string
            active:
              type: boolean
            createdAt:
              type: string
            updatedAt:
              type: string
            settings:
              type: object
              properties:
                font_size:
                  type: number
                volume:
                  type: number
                sound:
                  type: boolean
            preferences:
              type: object
              properties:
                language:
                  type: number
                color_icon:
                  type: array
                  items:
                    - type: string
                orientation_icon:
                  type: number
            statistics:
              type: object
              properties:
                game_time:
                  type: number
                tot_games:
                  type: number
                tot_points:
                  type: number
                max_points:
                  type: number
            country:
              type: object
              properties:
                _id:
                  type: string
                code:
                  type: string
                name:
                  type: string
                createdAt:
                  type: string
                updatedAt:
                  type: string
...
  requestBodies:
    login:
      description: Dati per effettuare il login
      content:
        application/json:
          schema:
            oneOf:
              - "$ref": "#/components/schemas/body_temporary"
              - "$ref": "#/components/schemas/body_credentials"
          example:
            uuid: 00000000-0000-0000-0000-000000000000
    registration:
      description: Dati per effettuare la registrazione del nuovo utente
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/body_registration"
          example:
            uuid: 00000000-0000-0000-0000-000000000000
            username: test
            email: test@gamery.com
            password: password
...
  responses:
    no_token:
      description: Nessun token di accesso trovato
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Token not found
    no_access:
      description: Accesso non consentito. Token o sessione non validi
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Token not valid
    base_response:
      description: Successo del completamento dell'azione
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: boolean
          example:
            data: true
    bad_request:
      description: Errore durante l'esecuzione dell'azione
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Descrizione dell'errore riscontrato
    login:
      description: Dati per l'accesso
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/response_login"
          example:
            access: token di accesso
            uuid: 00000000-0000-0000-0000-000000000000
    full_user:
      description: Dati completi dell'utente
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/response_full_user"
          example:
            data:
              statistics:
                game_time: 30
                tot_games: 2
                tot_points: 200
                max_points: 100
                best_placement: 1
              preferences:
                language: 1
                colors_icon:
                  - "#2AF39A"
                  - "#00A1F9"
                orientation_icon: 0
              settings:
                font_size: 14
                volume: 8
                sound: true
              _id: "000000000000000000000000"
              username: username
              uuid: 00000000-0000-0000-0000-000000000000
              active: true
              country:
                _id: "000000000000000000000000"
                code: af
                name: Afghanistan
                createdAt: "2022-01-01T00:00:00.0000Z"
                updatedAt: "2022-01-01T00:00:00.0000Z"
              createdAt: "2022-01-01T00:00:00.0000Z"
              updatedAt: "2022-01-01T00:00:00.0000Z"
...
paths:
  "/api/client/login":
    post:
      description: Effettua il login
      tags:
        - Users
      requestBody:
        "$ref": "#/components/requestBodies/login"
      responses:
        "200":
          "$ref": "#/components/responses/login"
        "400":
          "$ref": "#/components/responses/bad_request"
  "/api/client/register/temporary":
    get:
      description: Registra un nuovo utente temporaneo
      tags:
        - Users
      responses:
        "200":
          "$ref": "#/components/responses/login"
        "400":
          "$ref": "#/components/responses/bad_request"
  "/api/client/register":
    post:
      description: Registra un nuovo utente
      tags:
        - Users
      requestBody:
        "$ref": "#/components/requestBodies/registration"
      responses:
        "200":
          "$ref": "#/components/responses/login"
        "400":
          "$ref": "#/components/responses/bad_request"
  "/api/client/check":
    post:
      description: Controllo che il token sia valido
      tags:
        - Users
      security:
        - userAuth: []
      responses:
        "200":
          "$ref": "#/components/responses/base_response"
        "401":
          "$ref": "#/components/responses/no_token"
        "403":
          "$ref": "#/components/responses/no_access"
  "/api/client/logout":
    get:
      description: Effettua il logout
      tags:
        - Users
      security:
        - userAuth: []
      responses:
        "200":
          "$ref": "#/components/responses/base_response"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/no_token"
        "403":
          "$ref": "#/components/responses/no_access"
  "/api/client":
    get:
      description: Informazioni dell'utente
      tags:
        - Users
      security:
        - userAuth: []
      responses:
        "200":
          "$ref": "#/components/responses/full_user"
        "400":
          "$ref": "#/components/responses/bad_request"
        "401":
          "$ref": "#/components/responses/no_token"
        "403":
          "$ref": "#/components/responses/no_access"
  "/api/client/confirm":
    get:
      description: Conferma l'indirizzo email
      tags:
        - Users
      responses:
        "200":
          description: HTML con l'indicazione del successo dell'azione
        "400":
          description: HTML con la descrizione dell'errore
  "/api/client/change/password":
    get:
      description: Form per inviare la modifica della password
      tags:
        - Users
      parameters:
        - name: token
          description: Token inviato per email
          in: query
          required: true
      responses:
        "200":
          description: Pagina web per il ripristino della password
        "400":
          description: Errore durante l'esecuzione dell'azione
        "403":
          description: Accesso non consentito. Token non valido
    post:
      description: Resetta la password per l'utente
      tags:
        - Users
      parameters:
        - name: token
          description: Token inviato per email
          in: query
          required: true
      responses:
        "200":
          description: Successo dell'azione
        "400":
          description: Errore durante l'esecuzione dell'azione
        "403":
          description: Accesso non consentito. Token non valido
    put:
      description: Richiede un reset della password. Il link viene inviato per email
      tags:
        - Users
      responses:
        "200":
          description: Successo dell'azione
        "400":
          description: Errore durante l'esecuzione dell'azione
        "403":
          description: Accesso non consentito. Token non valido
  "/api/country":
    get:
      description: Ottiene tutti gli stati con le bandiere supportati
      tags:
        - Country
      responses:
        "200":
          description: Lista degli stati
        "403":
          description: Accesso negato
  "/api/country/client":
    get:
      description: Ottiene il country scelto dall'utente
      tags:
        - Country
      responses:
        "200":
          description: Il country selezionato dall'utente
        "400":
          description: Errore riscontrato nell'esecuzione
        "403":
          description: Accesso negato
    put:
      description: Aggiorna lo stato assegnato all'utente
      tags:
        - Country
      responses:
        "200":
          description: L'utente aggiornato
        "400":
          description: Errore riscontrato nell'esecuzione
        "403":
          description: Accesso negato
    delete:
      description: Rimuove il country selezionato dall'utente
      tags:
        - Country
      responses:
        "200":
          description: Il country selezionato dall'utente
        "400":
          description: Errore riscontrato nell'esecuzione
        "403":
          description: Accesso negato
...
  "/api/leaderboard/client":
    get:
      description: Ottiene la posizione in classifica dell'utente
      tags:
        - Leaderboard
      responses:
        "200":
          description: Posizione in classifica
        "400":
          description: Errore durante la generazione
        "403":
          description: Accesso non consentito
  "/api/leaderboard/client/{idOrUsername}":
    get:
      description:
        Ottiene la posizione in classifica di un utente in base all'id
        o al username
      tags:
        - Leaderboard
      responses:
        "200":
          description: Posizione in classifica
        "400":
          description: Errore durante la generazione
        "403":
          description: Accesso non consentito
...